@@ -4,7 +4,13 @@ import time
 import traceback
 import uuid
 from http import HTTPStatus
-from typing import Dict, List, Optional
+from typing import (
+    Callable,
+    Coroutine,
+    Dict,
+    List,
+    Optional,
+)
 
 from lnbits.core.services import create_invoice, fee_reserve, pay_invoice, service_fee
 from loguru import logger
@@ -12,11 +18,13 @@ from py_vapid import Vapid
 from pywebpush import WebPushException, webpush
 
 from lnbits.core.crud import (
-    delete_expired_invoices,
     delete_webpush_subscriptions,
     get_payments,
     get_standalone_payment,
+    update_payment_details,
+    update_payment_status,
 )
+from lnbits.core.models import Payment, PaymentState
 from lnbits.settings import settings
 from lnbits.wallets import get_funding_source
 
@@ -24,17 +32,13 @@ tasks: List[asyncio.Task] = []
 unique_tasks: Dict[str, asyncio.Task] = {}
 
 
-def create_task(coro):
+def create_task(coro: Coroutine) -> asyncio.Task:
     task = asyncio.create_task(coro)
     tasks.append(task)
     return task
 
 
-def create_permanent_task(func):
-    return create_task(catch_everything_and_restart(func))
-
-
-def create_unique_task(name: str, coro):
+def create_unique_task(name: str, coro: Coroutine) -> asyncio.Task:
     if unique_tasks.get(name):
         logger.warning(f"task `{name}` already exists, cancelling it")
         try:
@@ -46,11 +50,17 @@ def create_unique_task(name: str, coro):
     return task
 
 
-def create_permanent_unique_task(name: str, coro):
+def create_permanent_task(func: Callable[[], Coroutine]) -> asyncio.Task:
+    return create_task(catch_everything_and_restart(func))
+
+
+def create_permanent_unique_task(
+    name: str, coro: Callable[[], Coroutine]
+) -> asyncio.Task:
     return create_unique_task(name, catch_everything_and_restart(coro, name))
 
 
-def cancel_all_tasks():
+def cancel_all_tasks() -> None:
     for task in tasks:
         try:
             task.cancel()
@@ -63,9 +73,12 @@ def cancel_all_tasks():
             logger.warning(f"error while cancelling task `{name}`: {exc!s}")
 
 
-async def catch_everything_and_restart(func, name: str = "unnamed"):
+async def catch_everything_and_restart(
+    func: Callable[[], Coroutine],
+    name: str = "unnamed",
+) -> Coroutine:
     try:
-        await func()
+        return await func()
     except asyncio.CancelledError:
         raise  # because we must pass this up
     except Exception as exc:
@@ -73,7 +86,7 @@ async def catch_everything_and_restart(func, name: str = "unnamed"):
         logger.error(traceback.format_exc())
         logger.error("will restart the task in 5 seconds.")
         await asyncio.sleep(5)
-        await catch_everything_and_restart(func, name)
+        return await catch_everything_and_restart(func, name)
 
 
 invoice_listeners: Dict[str, asyncio.Queue] = {}
@@ -100,7 +113,7 @@ def register_invoice_listener(send_chan: asyncio.Queue, name: Optional[str] = No
 internal_invoice_queue: asyncio.Queue = asyncio.Queue(0)
 
 
-async def internal_invoice_listener():
+async def internal_invoice_listener() -> None:
     """
     internal_invoice_queue will be filled directly in core/services.py
     after the payment was deemed to be settled internally.
@@ -109,11 +122,11 @@ async def internal_invoice_listener():
     """
     while settings.lnbits_running:
         checking_id = await internal_invoice_queue.get()
-        logger.info("> got internal payment notification", checking_id)
-        create_task(invoice_callback_dispatcher(checking_id,True))
+        logger.info(f"got an internal payment notification {checking_id}")
+        await invoice_callback_dispatcher(checking_id, is_internal=True)
 
 
-async def invoice_listener():
+async def invoice_listener() -> None:
     """
     invoice_listener will collect all invoices that come directly
     from the backend wallet.
@@ -122,8 +135,23 @@ async def invoice_listener():
     """
     funding_source = get_funding_source()
     async for checking_id in funding_source.paid_invoices_stream():
-        logger.info("> got a payment notification", checking_id)
-        create_task(invoice_callback_dispatcher(checking_id,False))
+        logger.info(f"got a payment notification {checking_id}")
+        await invoice_callback_dispatcher(checking_id)
+
+
+def wait_for_paid_invoices(
+    invoice_listener_name: str,
+    func: Callable[[Payment], Coroutine],
+) -> Callable[[], Coroutine]:
+
+    async def wrapper() -> None:
+        invoice_queue: asyncio.Queue = asyncio.Queue()
+        register_invoice_listener(invoice_queue, invoice_listener_name)
+        while settings.lnbits_running:
+            payment = await invoice_queue.get()
+            await func(payment)
+
+    return wrapper
 
 
 async def check_pending_payments():
@@ -132,61 +160,70 @@ async def check_pending_payments():
     the backend and also to delete expired invoices. Incoming payments will be
     checked only once, outgoing pending payments will be checked regularly.
     """
-    outgoing = True
-    incoming = True
+    sleep_time = 60 * 30  # 30 minutes
 
     while settings.lnbits_running:
-        logger.info(
-            f"Task: checking all pending payments (incoming={incoming},"
-            f" outgoing={outgoing}) of last 15 days"
-        )
+        funding_source = get_funding_source()
+        if funding_source.__class__.__name__ == "VoidWallet":
+            logger.warning("Task: skipping pending check for VoidWallet")
+            await asyncio.sleep(sleep_time)
+            continue
         start_time = time.time()
         pending_payments = await get_payments(
             since=(int(time.time()) - 60 * 60 * 24 * 15),  # 15 days ago
             complete=False,
             pending=True,
-            outgoing=outgoing,
-            incoming=incoming,
             exclude_uncheckable=True,
         )
-        for payment in pending_payments:
-            await payment.check_status()
-            await asyncio.sleep(0.01)  # to avoid complete blocking
-
-        logger.info(
-            f"Task: pending check finished for {len(pending_payments)} payments"
-            f" (took {time.time() - start_time:0.3f} s)"
-        )
-        # we delete expired invoices once upon the first pending check
-        if incoming:
-            logger.debug("Task: deleting all expired invoices")
-            start_time = time.time()
-            await delete_expired_invoices()
+        count = len(pending_payments)
+        if count > 0:
+            logger.info(f"Task: checking {count} pending payments of last 15 days...")
+            for i, payment in enumerate(pending_payments):
+                status = await payment.check_status()
+                prefix = f"payment ({i+1} / {count})"
+                if status.failed:
+                    await update_payment_status(
+                        payment.checking_id, status=PaymentState.FAILED
+                    )
+                    logger.debug(f"{prefix} failed {payment.checking_id}")
+                elif status.success:
+                    await update_payment_details(
+                        checking_id=payment.checking_id,
+                        fee=status.fee_msat,
+                        preimage=status.preimage,
+                        status=PaymentState.SUCCESS,
+                    )
+                    logger.debug(f"{prefix} success {payment.checking_id}")
+                else:
+                    logger.debug(f"{prefix} pending {payment.checking_id}")
+                await asyncio.sleep(0.01)  # to avoid complete blocking
             logger.info(
-                "Task: expired invoice deletion finished (took"
-                f" {time.time() - start_time:0.3f} s)"
+                f"Task: pending check finished for {count} payments"
+                f" (took {time.time() - start_time:0.3f} s)"
             )
-
-        # after the first check we will only check outgoing, not incoming
-        # that will be handled by the global invoice listeners, hopefully
-        incoming = False
-
-        await asyncio.sleep(60 * 30)  # every 30 minutes
+        await asyncio.sleep(sleep_time)
 
 
-async def invoice_callback_dispatcher(checking_id: str,internal:bool):
+async def invoice_callback_dispatcher(checking_id: str, is_internal: bool = False):
     """
-    Takes incoming payments, sets pending=False, and dispatches them to
+    Takes an incoming payment, checks its status, and dispatches it to
     invoice_listeners from core and extensions.
     """
     payment = await get_standalone_payment(checking_id, incoming=True)
+    assert payment, "updated payment not found"
+    internal = "internal" if is_internal else ""
+    logger.success(f"{internal} invoice {checking_id} settled")
     if payment and payment.is_in:
-        logger.trace(
-            f"invoice listeners: sending invoice callback for payment {checking_id}"
+        status = await payment.check_status()
+        await update_payment_details(
+            checking_id=payment.checking_id,
+            fee=status.fee_msat,
+            preimage=status.preimage,
+            status=PaymentState.SUCCESS,
         )
         await payment.set_pending(False)
         # credit service fee wallet
-        service_fee_msat = service_fee(payment.amount, internal=internal)
+        service_fee_msat = service_fee(payment.amount, internal=is_internal)
         print('>config',settings.lnbits_service_fee_wallet,service_fee_msat)
         if settings.lnbits_service_fee_wallet and service_fee_msat and service_fee_msat>=1000 and payment.wallet_id!=settings.lnbits_service_fee_wallet:
             print('service-fee-msat',service_fee_msat)
@@ -205,6 +242,7 @@ async def invoice_callback_dispatcher(checking_id: str,internal:bool):
                 )
               
         
+       
         for name, send_chan in invoice_listeners.items():
             logger.trace(f"invoice listeners: sending to `{name}`")
             await send_chan.put(payment)
@@ -225,8 +263,11 @@ async def send_push_notification(subscription, title, body, url=""):
             {"aud": "", "sub": "mailto:alan@lnbits.com"},
         )
     except WebPushException as e:
-        if e.response.status_code == HTTPStatus.GONE:
+        if e.response and e.response.status_code == HTTPStatus.GONE:
             # cleanup unsubscribed or expired push subscriptions
             await delete_webpush_subscriptions(subscription.endpoint)
         else:
-            logger.error(f"failed sending push notification: {e.response.text}")
+            logger.error(
+                f"failed sending push notification: "
+                f"{e.response.text if e.response else e}"
+            )
