@@ -55,8 +55,8 @@ cursor.execute(
     """
     INSERT INTO apipayments
       (wallet, checking_id, bolt11, hash, preimage,
-       amount, pending, memo, fee, extra, webhook, expiry)
-    VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)
+       amount, status, memo, fee, extra, webhook, expiry, pending)
+    VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)
     """,
     (
         wallet_id,
@@ -65,12 +65,13 @@ cursor.execute(
         "test_admin_internal",
         None,
         amount * 1000,
-        False,
+        "success",
         "test_admin_internal",
         0,
         None,
         "",
         expiration_date,
+        False,  # TODO: remove this in next release
     ),
 )
 
